WIP refactor

+++ src/song.hpp
@@ -95,7 +95,7 @@ struct Effect {
 
 
 struct Song {
-    using Block = std::array<int, CHANNEL_COUNT>;
+    using Block = std::array<uint8_t, CHANNEL_COUNT>;
 
     uint8_t tempo; // 4 to F
     uint8_t swing; // 0 to 4
